backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
supabase==2.0.0
bcrypt==4.1.0
PyJWT==2.8.0
requests==2.31.0
